bump libnfs to 1.3

+++ xbmc/filesystem/DllLibNfs.h
@@ -44,8 +44,8 @@ public:
   virtual void   free_nfs_srvr_list(struct nfs_server_list *srv)=0;
   virtual struct nfs_context *nfs_init_context(void)=0;
   virtual void nfs_destroy_context(struct nfs_context *nfs)=0;  
-  virtual size_t nfs_get_readmax(struct nfs_context *nfs)=0;  
-  virtual size_t nfs_get_writemax(struct nfs_context *nfs)=0;
+  virtual uint64_t nfs_get_readmax(struct nfs_context *nfs)=0;  
+  virtual uint64_t nfs_get_writemax(struct nfs_context *nfs)=0;
   virtual char *nfs_get_error(struct nfs_context *nfs)=0;  
   virtual int nfs_close(struct nfs_context *nfs,     struct nfsfh *nfsfh)=0;
   virtual int nfs_fsync(struct nfs_context *nfs,     struct nfsfh *nfsfh)=0;  
@@ -57,8 +57,8 @@ public:
   virtual int nfs_mount(struct nfs_context *nfs,     const char *server,   const char *exportname)=0;
   virtual int nfs_stat(struct nfs_context *nfs,      const char *path,     struct stat *st)=0;
   virtual int nfs_fstat(struct nfs_context *nfs,     struct nfsfh *nfsfh,  struct stat *st)=0;
-  virtual int nfs_truncate(struct nfs_context *nfs,  const char *path,     off64_t length)=0;
-  virtual int nfs_ftruncate(struct nfs_context *nfs, struct nfsfh *nfsfh,  off64_t length)=0;
+  virtual int nfs_truncate(struct nfs_context *nfs,  const char *path,     uint64_t length)=0;
+  virtual int nfs_ftruncate(struct nfs_context *nfs, struct nfsfh *nfsfh,  uint64_t length)=0;
   virtual int nfs_opendir(struct nfs_context *nfs,   const char *path,     struct nfsdir **nfsdir)=0;
   virtual int nfs_statvfs(struct nfs_context *nfs,   const char *path,     struct statvfs *svfs)=0;
   virtual int nfs_chmod(struct nfs_context *nfs,     const char *path,     int mode)=0;
@@ -73,12 +73,12 @@ public:
   virtual int nfs_chown(struct nfs_context *nfs,     const char *path,     int uid,      int gid)=0;
   virtual int nfs_fchown(struct nfs_context *nfs,    struct nfsfh *nfsfh,  int uid,      int gid)=0;
   virtual int nfs_open(struct nfs_context *nfs,      const char *path,     int mode,     struct nfsfh **nfsfh)=0;  
-  virtual int nfs_read(struct nfs_context *nfs,      struct nfsfh *nfsfh,  size_t count, char *buf)=0;  
-  virtual int nfs_write(struct nfs_context *nfs,     struct nfsfh *nfsfh,  size_t count, char *buf)=0;
+  virtual int nfs_read(struct nfs_context *nfs,      struct nfsfh *nfsfh,  uint64_t count, char *buf)=0;  
+  virtual int nfs_write(struct nfs_context *nfs,     struct nfsfh *nfsfh,  uint64_t count, char *buf)=0;
   virtual int nfs_creat(struct nfs_context *nfs,     const char *path,     int mode,     struct nfsfh **nfsfh)=0;  
-  virtual int nfs_pread(struct nfs_context *nfs,     struct nfsfh *nfsfh,  off64_t offset, size_t count, char *buf)=0;
-  virtual int nfs_pwrite(struct nfs_context *nfs,    struct nfsfh *nfsfh,  off64_t offset, size_t count, char *buf)=0;
-  virtual int nfs_lseek(struct nfs_context *nfs,     struct nfsfh *nfsfh,  off64_t offset, int whence,   off64_t *current_offset)=0;
+  virtual int nfs_pread(struct nfs_context *nfs,     struct nfsfh *nfsfh,  uint64_t offset, uint64_t count, char *buf)=0;
+  virtual int nfs_pwrite(struct nfs_context *nfs,    struct nfsfh *nfsfh,  uint64_t offset, uint64_t count, char *buf)=0;
+  virtual int nfs_lseek(struct nfs_context *nfs,     struct nfsfh *nfsfh,  uint64_t offset, int whence,   uint64_t *current_offset)=0;
 };
 
 class DllLibNfs : public DllDynamic, DllLibNfsInterface
@@ -90,8 +90,8 @@ class DllLibNfs : public DllDynamic, DllLibNfsInterface
   DEFINE_METHOD1(struct exportnode *, mount_getexports,     (const char *p1))
   DEFINE_METHOD1(void,    mount_free_export_list,           (struct exportnode *p1))
   DEFINE_METHOD1(void,    nfs_destroy_context,              (struct nfs_context *p1))
-  DEFINE_METHOD1(size_t,  nfs_get_readmax,                  (struct nfs_context *p1))
-  DEFINE_METHOD1(size_t,  nfs_get_writemax,                 (struct nfs_context *p1)) 
+  DEFINE_METHOD1(uint64_t,  nfs_get_readmax,                  (struct nfs_context *p1))
+  DEFINE_METHOD1(uint64_t,  nfs_get_writemax,                 (struct nfs_context *p1)) 
   DEFINE_METHOD1(char *,  nfs_get_error,                    (struct nfs_context *p1))    
   DEFINE_METHOD2(struct nfsdirent *, nfs_readdir,           (struct nfs_context *p1, struct nfsdir *p2))
   DEFINE_METHOD2(int, nfs_fsync,     (struct nfs_context *p1, struct nfsfh *p2))
@@ -103,8 +103,8 @@ class DllLibNfs : public DllDynamic, DllLibNfsInterface
   DEFINE_METHOD3(int, nfs_mount,     (struct nfs_context *p1, const char *p2,    const char *p3))
   DEFINE_METHOD3(int, nfs_stat,      (struct nfs_context *p1, const char *p2,    struct stat *p3))
   DEFINE_METHOD3(int, nfs_fstat,     (struct nfs_context *p1, struct nfsfh *p2,  struct stat *p3))
-  DEFINE_METHOD3(int, nfs_truncate,  (struct nfs_context *p1, const char *p2,    off64_t p3))
-  DEFINE_METHOD3(int, nfs_ftruncate, (struct nfs_context *p1, struct nfsfh *p2,  off64_t p3))
+  DEFINE_METHOD3(int, nfs_truncate,  (struct nfs_context *p1, const char *p2,    uint64_t p3))
+  DEFINE_METHOD3(int, nfs_ftruncate, (struct nfs_context *p1, struct nfsfh *p2,  uint64_t p3))
   DEFINE_METHOD3(int, nfs_opendir,   (struct nfs_context *p1, const char *p2,    struct nfsdir **p3))
   DEFINE_METHOD3(int, nfs_statvfs,   (struct nfs_context *p1, const char *p2,    struct statvfs *p3))
   DEFINE_METHOD3(int, nfs_chmod,     (struct nfs_context *p1, const char *p2,    int p3))
@@ -116,15 +116,15 @@ class DllLibNfs : public DllDynamic, DllLibNfsInterface
   DEFINE_METHOD3(int, nfs_rename,    (struct nfs_context *p1, const char *p2,    const char *p3))
   DEFINE_METHOD3(int, nfs_link,      (struct nfs_context *p1, const char *p2,    const char *p3))  
   DEFINE_METHOD4(int, nfs_open,      (struct nfs_context *p1, const char *p2,    int p3,     struct nfsfh **p4))
-  DEFINE_METHOD4(int, nfs_read,      (struct nfs_context *p1, struct nfsfh *p2,  size_t p3,  char *p4))
-  DEFINE_METHOD4(int, nfs_write,     (struct nfs_context *p1, struct nfsfh *p2,  size_t p3,  char *p4))
+  DEFINE_METHOD4(int, nfs_read,      (struct nfs_context *p1, struct nfsfh *p2,  uint64_t p3,  char *p4))
+  DEFINE_METHOD4(int, nfs_write,     (struct nfs_context *p1, struct nfsfh *p2,  uint64_t p3,  char *p4))
   DEFINE_METHOD4(int, nfs_creat,     (struct nfs_context *p1, const char *p2,    int p3,     struct nfsfh **p4))
   DEFINE_METHOD4(int, nfs_readlink,  (struct nfs_context *p1, const char *p2,    char *p3,   int p4))
   DEFINE_METHOD4(int, nfs_chown,     (struct nfs_context *p1, const char *p2,    int p3,     int p4))
   DEFINE_METHOD4(int, nfs_fchown,    (struct nfs_context *p1, struct nfsfh *p2,  int p3,     int p4))
-  DEFINE_METHOD5(int, nfs_pread,     (struct nfs_context *p1, struct nfsfh *p2,  off64_t p3,   size_t p4,  char *p5))
-  DEFINE_METHOD5(int, nfs_pwrite,    (struct nfs_context *p1, struct nfsfh *p2,  off64_t p3,   size_t p4,  char *p5))
-  DEFINE_METHOD5(int, nfs_lseek,     (struct nfs_context *p1, struct nfsfh *p2,  off64_t p3,   int p4,     off64_t *p5))
+  DEFINE_METHOD5(int, nfs_pread,     (struct nfs_context *p1, struct nfsfh *p2,  uint64_t p3,   uint64_t p4,  char *p5))
+  DEFINE_METHOD5(int, nfs_pwrite,    (struct nfs_context *p1, struct nfsfh *p2,  uint64_t p3,   uint64_t p4,  char *p5))
+  DEFINE_METHOD5(int, nfs_lseek,     (struct nfs_context *p1, struct nfsfh *p2,  uint64_t p3,   int p4,     uint64_t *p5))
